Finished parsing continuous
